fix: move `QuikPy` lib
