+++ Examples/01_Connect.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from QuikPy import QuikPy  # Работа с QUIK из Python через LUA скрипты QuikSharp
+from QUIK.QuikPy import QuikPy  # Работа с QUIK из Python через LUA скрипты QuikSharp
 
 
 if __name__ == '__main__':  # Точка входа при запуске этого скрипта
