--- tickers
@@ -1,1 +1,0 @@
-{'first_curr_qty_scale': 0, 'step_price_currency': '', 'stock_name': '', 'first_currcode': '', 'code': 'HEAD', 'regnumber': '1-01-16755-A', 'name': 'МКПАО Хэдхантер', 'trade_currency': '', 'base_active_seccode': '', 'buybackprice': 0.0, 'face_unit': 'SUR', 'isin_code': 'RU000A107662', 'scale': 0, 'bsid': '', 'class_code': 'QJSIM', 'face_value': 1.0, 'couponvalue': 0.0, 'stock_code': '', 'exp_date': 0, 'cusip_code': '', 'mat_date': 0, 'accruedint': 0.0, 'qty_scale': 0, 'settlecode': '', 'second_curr_qty_scale': 0, 'buybackdate': 0, 'sec_code': 'HEAD', 'base_active_classcode': '', 'option_strike': 0.0, 'class_name': 'Акции 1-го уровня (эмулятор)', 'yieldatprevwaprice': 0.0, 'min_price_step': 1.0, 'sedol_code': '', 'qty_multiplier': 1, 'settle_date': 0, 'nextcoupon': 0, 'couponperiod': 0, 'list_level': 0, 'short_name': 'Хэдхантер', 'ric_code': '', 'second_currcode': '', 'cfi_code': '', 'lot_size': 1}
