--- ticket
@@ -1,1 +1,0 @@
-{'code': 'BRQ8', 'couponvalue': 0.0, 'base_active_seccode': 'BR', 'stock_name': '', 'list_level': 0, 'trade_currency': 'SUR', 'buybackprice': 0.0, 'short_name': 'BRQ8', 'bsid': '', 'sec_code': 'BRQ8', 'regnumber': '', 'accruedint': 0.0, 'stock_code': '', 'second_curr_qty_scale': 0, 'exp_date': 0, 'cusip_code': '', 'settle_date': 0, 'name': 'BR-8.18', 'settlecode': '', 'first_curr_qty_scale': 0, 'class_name': 'МБ Срочный рынок: Фьючерсы', 'class_code': 'SPBFUT', 'step_price_currency': 'USD', 'qty_scale': 0, 'buybackdate': 0, 'first_currcode': '', 'yieldatprevwaprice': 0.0, 'lot_size': 10, 'scale': 2, 'cfi_code': '', 'couponperiod': 0, 'isin_code': '', 'sedol_code': '', 'mat_date': 20180801, 'base_active_classcode': '', 'second_currcode': '', 'nextcoupon': 0, 'option_strike': 0.0, 'ric_code': '', 'min_price_step': 0.01, 'face_value': 0.0, 'face_unit': 'SUR', 'qty_multiplier': 1}
